requirements.txt
pytest
matplotlib
numpy
trimesh
pyserial
ultralytics
opencv-python
ezdxf
click
fastapi
uvicorn
pyyaml

mkdocs
mkdocstrings[python]
websockets
requests
psutil
shapely
